move createBrowseLinks to api from collections module to avoid circular dependency
