+++ search/tests/convert/collections.spec.js
@@ -2,8 +2,7 @@ const {
   cmrCollSpatialToExtents,
   stacSearchWithCurrentParams,
   cmrGranuleSearchWithCurrentParams,
-  cmrCollToWFSColl,
-  createBrowseLinks
+  cmrCollToWFSColl
 } = require('../../lib/convert/collections');
 const axios = require('axios');
 const { WHOLE_WORLD_BBOX } = require('../../lib/convert');
@@ -318,15 +317,6 @@ describe('collections', () => {
         short_name: 'id-LPDAAC',
         stac_version: settings.stac.version,
         license: 'not-provided'
-      });
-    });
-
-    it('should return browse links', async () => {
-      const links = await createBrowseLinks(event, cmrColl.data_center, cmrColl.dataset_id);
-      expect(links.length).toEqual(5);
-      [1, 2, 3, 4, 5].forEach(x => {
-        expect(links[x - 1].href)
-          .toEqual(`http://example.com/stac/LPDAAC/collections/datasetId/200${x}`);
       });
     });
   });
